requirements.txt
Flask==1.1.2
Flask_MySQLdb==0.2.0
